ci: run ./cmd/... on all native-builds legs (Issue #3470) (#3477)
* ci: run ./cmd/... on all native-builds legs (Issue #3470)

Add ./cmd/... to the "Run Unit Tests" step of native-builds in
cross-platform-build.yml for all three platforms (Linux, macOS, Windows).

The Windows leg is the critical gap: Windows-only SCM and service-management
code in cmd/cfgms-steward-launcher, cmd/steward/service, and
cmd/controller/service is //go:build windows — invisible to the Linux
unit-tests job. This is the only job that runs on a Windows runner, so
those tests executed in no CI job before this change.

Linux and macOS are included for consistency; their coverage is already
provided by the unit-tests job via make test, so ./cmd/... there is
additive but not the primary driver.

The 8m timeout comment is updated to account for SCM / WMI calls in
./cmd/... on Windows. Local timing on a Linux dev container shows
./cmd/... completes in ~25s; at the documented ~7× overhead on 2–3 vCPU
hosted runners that is well within the 8m wall.

Fixes #3470


* fix: address PR #3477 review comments (Issue #3470)

Fix the Windows-only StartType assertion in
TestRepairServiceRegistration_RecreatesService: mgr.StartAutomatic is an
untyped constant that defaults to int when boxed for assert.Equal, while
mgr.Config.StartType is uint32 — cast to uint32 so the comparison actually
compares equal values instead of failing on type mismatch
(expected: int(2) actual: uint32(0x2)).

The other two failures the acceptance review flagged
(TestDNACollectorAdapter_CollectFragments_ForwardsToModuleSource and
_UnionsHostAndModuleFragments in cmd/steward) are pre-existing,
Windows-only timing failures unrelated to this workflow change — they pass
cleanly on Linux and were already flagged as a Windows-triage risk in
#3470's own Constraints section. Filed and linked follow-up story #3483
(under epic #2911) rather than folding an unrelated product fix into this
CI workflow story, per #3470's explicit Out of Scope guidance.


* ci: run ./cmd/... with -v on the Windows leg (Issue #3470)

The Windows leg's "Run Unit Tests" step ran every package non-verbose, so the
job log carried one "ok <package>" line per package and no per-test result
lines at all. The Windows-only SCM tests in cmd/cfgms-steward-launcher skip
themselves when the process is not elevated (requireSCM calls t.Skipf when
mgr.Connect fails), and a non-verbose log cannot distinguish an SCM test that
passed from one that skipped — both roll up into the same "ok" line.

Split the Windows invocation so ./cmd/... runs with -v while ./pkg/... and
./features/... stay non-verbose. -v is scoped to ./cmd/... deliberately: those
are the packages this job exists to cover, and enabling it fleet-wide would
bury the SCM results under every other package's per-test output.

No test was added, removed, skipped, or build-tagged. The Linux/macOS branch of
the step is unchanged.

Fixes #3470


---------

+++ cmd/cfgms-steward-launcher/service_check_windows_test.go
@@ -130,7 +130,7 @@ func TestRepairServiceRegistration_RecreatesService(t *testing.T) {
 
 	cfg, err := s.Config()
 	require.NoError(t, err)
-	assert.Equal(t, mgr.StartAutomatic, cfg.StartType, "start type must match install")
+	assert.Equal(t, uint32(mgr.StartAutomatic), cfg.StartType, "start type must match install")
 	assert.Equal(t, launcherServiceDisplayName, cfg.DisplayName)
 	assert.Equal(t, launcherServiceDescription, cfg.Description)
 	assert.Contains(t, cfg.BinaryPathName, testRepairExePath,
